Added the concept of closure
I had used some elaborate methods while adding items and later on removing them: I added counter to create unique Ids for each new item being created and also assigned the same Id to the delete button being created.

This measure was implemented thinking that while removing a particular <li> item, I will need a mechanism to point to that item. So, when the delete <button> associated with Id-2 <li> item is clicked, the removal function would check the matching <li> item using the Id and then remove it. (Of course, in the later version of the code, I used closest() method, so in a way, this Id addition was obsolete.)

But I learned that whenever we create an eventListener for removal as in our case and assign the removal parameters-including the parents etc right away, this is stored as a context and it is called closure.

In this document, I have copied the whole conversation when I compared the solution code with mine.